+++ src/applications/model/onoff-application.cc
@@ -198,7 +198,7 @@ OnOffApplication::ScheduleNextTx()
         Time nextTime(
             Seconds(bits / static_cast<double>(m_cbrRate.GetBitRate()))); // Time till next packet
         NS_LOG_LOGIC("nextTime = " << nextTime.As(Time::S));
-        m_sendEvent = Simulator::Schedule(nextTime, &OnOffApplication::SendPacket, this);
+        m_sendEvent = Simulator::Schedule(nextTime, &OnOffApplication::TransmitPacket, this);
     }
     else
     {
@@ -229,57 +229,14 @@ OnOffApplication::ScheduleStopEvent()
 }
 
 void
-OnOffApplication::SendPacket()
+OnOffApplication::TransmitPacket()
 {
     NS_LOG_FUNCTION(this);
 
     NS_ASSERT(m_sendEvent.IsExpired());
 
-    Ptr<Packet> packet;
-    if (m_unsentPacket)
-    {
-        packet = m_unsentPacket;
-    }
-    else if (m_enableSeqTsSizeHeader)
-    {
-        uint32_t headerSize{0};
-        SeqTsSizeHeader seqTsSizeHdr;
-        SeqTsHeader seqTsHdr;
-        if (m_socket->GetSocketType() == Socket::NS3_SOCK_STREAM)
-        {
-            seqTsSizeHdr.SetSeq(m_seq++);
-            seqTsSizeHdr.SetSize(m_pktSize);
-            headerSize = seqTsSizeHdr.GetSerializedSize();
-        }
-        else if (m_socket->GetSocketType() == Socket::NS3_SOCK_DGRAM)
-        {
-            seqTsHdr.SetSeq(m_seq++);
-            headerSize = seqTsHdr.GetSerializedSize();
-        }
-        NS_ABORT_IF(m_pktSize < headerSize);
-        packet = Create<Packet>(m_pktSize - headerSize);
-        // Trace before adding header, for consistency with PacketSink
-        Address from;
-        Address to;
-        m_socket->GetSockName(from);
-        m_socket->GetPeerName(to);
-        if (m_socket->GetSocketType() == Socket::NS3_SOCK_STREAM)
-        {
-            m_txTraceWithSeqTsSize(packet, from, to, seqTsSizeHdr);
-            packet->AddHeader(seqTsSizeHdr);
-        }
-        else if (m_socket->GetSocketType() == Socket::NS3_SOCK_DGRAM)
-        {
-            m_txTraceWithSeqTs(packet, from, to, seqTsHdr);
-            packet->AddHeader(seqTsHdr);
-        }
-    }
-    else
-    {
-        packet = Create<Packet>(m_pktSize);
-    }
-
-    int actual = m_socket->Send(packet);
+    auto packet = m_unsentPacket ? m_unsentPacket : CreatePacket(m_pktSize);
+    int actual = SendPacket(packet);
     if ((unsigned)actual == m_pktSize)
     {
         m_txTrace(packet);