+++ src/applications/model/onoff-application.h
@@ -120,7 +120,7 @@ class OnOffApplication : public SourceApplication
     /**
      * @brief Send a packet
      */
-    void SendPacket();
+    void TransmitPacket();
     /**
      * @brief Schedule the next packet transmission
      */
@@ -145,7 +145,6 @@ class OnOffApplication : public SourceApplication
     uint64_t m_totBytes{0};              //!< Total bytes sent so far
     EventId m_startStopEvent;            //!< Event id for next start or stop event
     EventId m_sendEvent;                 //!< Event id of pending "send packet" event
-    uint32_t m_seq{0};                   //!< Sequence
     Ptr<Packet> m_unsentPacket;          //!< Unsent packet cached for future attempt
 
     /// Callbacks for tracing the packet Tx events, includes source and destination addresses